ResidentService tests

+++ src/main/java/com/senai/vila/controller/service/ResidentService.java
@@ -23,7 +23,24 @@ public class ResidentService {
         this.residentRepository = residentRepository;
     }
 
+    public ResidentDto createResident(ResidentDto residentDto) throws ResidentException, CpfException {
+        if (residentDto == null) {
+            throw new ResidentException("Habitante não pode ser nulo");
+        }
+        String cpf = residentDto.getCpf();
+        residentDto.setCpf(CpfValidator.validateCpf(cpf));
+
+        if (residentRepository.findByCpf(residentDto.getCpf()).isPresent()) {
+            throw new ResidentException("Já existe um habitante com esse cpf cadastrado!");
+        }
+        Resident residentDao = residentRepository.save(new Resident(residentDto));
+        return residentDao.convertToDto();
+    }
+
     public Resident getResidentByEmail(String email) throws ResidentException {
+        if (email == null) {
+            throw new ResidentException("Email não pode ser nulo");
+        }
         Optional<Resident> habitante = residentRepository.findByEmail(email);
         if (habitante.isPresent()) {
             return habitante.get();
@@ -44,26 +61,26 @@ public class ResidentService {
         throw new ResidentException("Não existe habitante com esse id");
     }
 
-    public ResidentDto createResident(ResidentDto residentDto) throws ResidentException, CpfException {
-        String cpf = residentDto.getCpf();
-        residentDto.setCpf(CpfValidator.validateCpf(cpf));
-
-        if (residentRepository.findByCpf(residentDto.getCpf()).isPresent()) {
-            throw new ResidentException("Já existe um habitante com esse cpf cadastrado!");
+    public void deleteResident(Long id) throws ResidentException {
+        if (id == null) {
+            throw new ResidentException("Id não pode ser nulo");
         }
-        Resident residentDao = residentRepository.save(new Resident(residentDto));
-        return residentDao.convertToDto();
-    }
-
-    public void deleteResident(Long id) {
         residentRepository.deleteById(id);
     }
 
-    public Optional<Resident> getResidentByCpf(String cpf) {
-        return residentRepository.findByCpf(cpf);
+    public ResidentDto getResidentByCpf(String cpf) throws ResidentException, CpfException {
+        if (cpf == null) {
+            throw new ResidentException("Cpf não pode ser nulo");
+        }
+        cpf = CpfValidator.validateCpf(cpf);
+        Optional<Resident> residentOptional = residentRepository.findByCpf(cpf);
+        if (residentOptional.isPresent()) {
+            return residentOptional.get().convertToDto();
+        }
+        throw new ResidentException("Habitante não encontrado");
     }
 
-    public List<ResidentDto> listByBirthMonth(Integer month) {
+    public List<ResidentDto> listByBirthMonth(Integer month) throws ResidentException {
         if (month == null) {
             throw new IllegalArgumentException("Mês não pode ser nulo");
         }
@@ -81,10 +98,10 @@ public class ResidentService {
                 return residentDto;
             }).collect(Collectors.toList());
         }
-        return null;
+        throw new ResidentException("Não existe habitantes com esse mês de nascimento");
     }
 
-    public List<ResidentDto> listByFirstName(String firstName) {
+    public List<ResidentDto> listByFirstName(String firstName) throws ResidentException {
         if (firstName == null) {
             throw new IllegalArgumentException("Nome não pode ser nulo");
         }
@@ -96,10 +113,10 @@ public class ResidentService {
                 return residentDto;
             }).collect(Collectors.toList());
         }
-        return null;
+        throw new ResidentException("Não existe habitantes com esse primeiro nome");
     }
 
-    public List<ResidentDto> listByLastName(String lastName) {
+    public List<ResidentDto> listByLastName(String lastName) throws ResidentException {
         if (lastName == null) {
             throw new IllegalArgumentException("Sobrenome não pode ser nulo");
         }
@@ -111,10 +128,10 @@ public class ResidentService {
                 return residentDto;
             }).collect(Collectors.toList());
         }
-        return null;
+        throw new ResidentException("Não existe habitantes com esse sobrenome");
     }
 
-    public List<ResidentDto> listByAge(Integer age) {
+    public List<ResidentDto> listByAge(Integer age) throws ResidentException {
         if (age == null) {
             throw new IllegalArgumentException("Idade não pode ser nula");
         }
@@ -122,7 +139,7 @@ public class ResidentService {
             throw new IllegalArgumentException("Idade não pode ser negativa");
         }
         LocalDate dateBirth = LocalDate.now().minusYears(age);
-        Optional<List<Resident>> residents = residentRepository.findByAge(dateBirth);
+        Optional<List<Resident>> residents = residentRepository.findByAge(age);
         if (residents.isPresent()) {
             return residents.get().stream().map(resident -> {
                 ResidentDto residentDto = resident.convertToDto();
@@ -130,7 +147,7 @@ public class ResidentService {
                 return residentDto;
             }).collect(Collectors.toList());
         }
-        return null;
+        throw new ResidentException("Não existe habitantes com essa idade");
     }
 
     public Double getTotalRentResidents() {
